// src/review/review.enum.ts
export enum SatisfactionLevel {
  SATISFIED = 'SATISFIED',
  NEEDS_IMPROVEMENT = 'NEEDS_IMPROVEMENT',
  UNSATISFIED = 'UNSATISFIED',
}

export enum IterationStatus {
  PENDING_REVIEW = 'PENDING_REVIEW',
  COMPLETED = 'COMPLETED',
  NEEDS_ITERATION = 'NEEDS_ITERATION',
  MATCHED = 'MATCHED',
  MISSING = 'MISSING',
  MODIFIED = 'MODIFIED',
  APPROVED = 'APPROVED',
  IN_REVIEW = 'IN_REVIEW',
  PENDING = 'PENDING',
}